+++ app/org/nlp4l/framework/controllers/JobController.scala
@@ -23,7 +23,6 @@ import java.io.PrintWriter
 import java.nio.file.Files
 import java.nio.file.Paths
 import java.util.UUID
-
 import scala.collection.JavaConversions.asScalaBuffer
 import scala.concurrent.Await
 import scala.concurrent.ExecutionContext.Implicits.global
@@ -31,9 +30,7 @@ import scala.concurrent.Future
 import scala.concurrent.TimeoutException
 import scala.util.Failure
 import scala.util.Success
-
 import com.google.inject.name.Named
-
 import akka.actor.ActorRef
 import javax.inject.Inject
 import play.api.libs.json.JsValue.jsValueToJsLookup
@@ -41,29 +38,28 @@ import play.api.libs.json.Json
 import play.api.libs.json.Json.toJsFieldJsValueWrapper
 import play.api.mvc.Action
 import play.api.mvc.Controller
-
 import org.joda.time.DateTime
 import org.nlp4l.framework.dao.JobDAO
 import org.nlp4l.framework.dao.RunDAO
-import org.nlp4l.framework.models.ActionResult
 import org.nlp4l.framework.models.Cell
 import org.nlp4l.framework.models.CellAttribute
 import org.nlp4l.framework.models.CellType
-import org.nlp4l.framework.models.DbModels.fWActionResultWrites
-import org.nlp4l.framework.models.DbModels.fWJobStatusWrites
-import org.nlp4l.framework.models.DbModels.fWJobWrites
-import org.nlp4l.framework.models.DbModels.fWRecordWithAttrubuteWrites
+import org.nlp4l.framework.buildin.DbModels.fWActionResultWrites
+import org.nlp4l.framework.buildin.DbModels.fWJobStatusWrites
+import org.nlp4l.framework.buildin.DbModels.fWJobWrites
+import org.nlp4l.framework.buildin.DbModels.fWRecordWithAttrubuteWrites
 import org.nlp4l.framework.models.Dictionary
 import org.nlp4l.framework.models.DictionaryAttribute
-import org.nlp4l.framework.models.Job
-import org.nlp4l.framework.models.JobMessage
 import org.nlp4l.framework.models.Record
 import org.nlp4l.framework.models.RecordWithAttrbute
-import org.nlp4l.framework.models.Replay
-import org.nlp4l.framework.processors.DeployerChain
+import org.nlp4l.framework.buildin.DeployerChain
 import org.nlp4l.framework.processors.ProcessorChain2
 import org.nlp4l.framework.processors.ProcessorChain2Builder
-import org.nlp4l.framework.processors.ValidatorChain
+import org.nlp4l.framework.buildin.ValidatorChain
+import org.nlp4l.framework.buildin.Replay
+import org.nlp4l.framework.buildin.JobMessage
+import org.nlp4l.framework.buildin.ActionResult
+import org.nlp4l.framework.buildin.Job
