+++ app/org/nlp4l/framework/views/jobresult.scala.html
@@ -1,4 +1,4 @@
-@import org.nlp4l.framework.models.Job
+@import org.nlp4l.framework.buildin.Job
 @(job: Job, jobId: Int, runId: Int, runIdList: Seq[Int], listTable: String, addForm: String, editForm: String)
 
 @headerAddon = {
